IEREP-51: Added all DoD renders
