+++ src/pages/subPages/DiagnosisOfDeath/renders/DODSixth/DODSixth.component.jsx
@@ -1,0 +1,136 @@
+// Import: Dependencies
+import React, { useState, useEffect } from "react";
+import { diagnosisOfDeath } from "../../../../../utils/axios";
+
+// Import: Elements
+import {
+  ColumnOne,
+  ColumnTwo,
+  Grid,
+  Heading,
+  Render,
+  Section,
+} from "./DODSixth.elements";
+
+// Import: Components
+import {
+  FieldData,
+  HeadingPrimary,
+  Loader,
+  ReportContainer,
+} from "../../../../../components";
+
+// Render: DODSixth
+export default function DODSixth() {
+  // State = loading, sectionData
+  const [loading, setLoading] = useState(true);
+  const [sectionData, setSectionData] = useState([]);
+
+  // Patient Master_ePR_ID
+  const patientId = "282e3dbb-a766-4185-8343-45e4e12d3587";
+
+  // Fetch Patient data from OneResponse API
+  function getPatientData() {
+    setLoading(true);
+
+    diagnosisOfDeath
+      .get(`/${patientId}`, {})
+      .then((res) => {
+        const data = res.data;
+        setSectionData(data);
+        setLoading(false);
+      })
+      .catch((error) => {
+        setLoading(false);
+        console.log(error);
+      });
+  }
+
+  // Run getPatientData on load
+  useEffect(() => {
+    getPatientData();
+  }, []);
+
+  //#region sectionRender = Expected Death... Report
+  const sectionRender = sectionData.map((patient) => (
+    <React.Fragment key={patient.id}>
+      {/* Expected Death */}
+      <Section>
+        <Heading>
+          <HeadingPrimary
+            icon="fas fa-sticky-note"
+            text="E. Expected Death Due to Illness or Pre-Existing Condition"
+            padding="0.6rem"
+          />
+        </Heading>
+
+        <ReportContainer>
+          <Render>
+            <Grid>
+              <ColumnOne>
+                <FieldData
+                  field="Expected Death due to a terminal illness"
+                  data={
+                    patient.DoD_Terminal ? patient.DoD_Terminal : "Not recorded"
+                  }
+                />
+              </ColumnOne>
+
+              <ColumnTwo>
+                <FieldData
+                  field="Explanation"
+                  data={
+                    patient.Terminal_Illness
+                      ? patient.Terminal_Illness
+                      : "Not recorded"
+                  }
+                />
+              </ColumnTwo>
+            </Grid>
+          </Render>
+        </ReportContainer>
+      </Section>
+    </React.Fragment>
+  ));
+  //#endregion /sectionRender = Expected Death... Report
+
+  return (
+    <>
+      {loading ? (
+        <Section>
+          <Heading>
+            <HeadingPrimary
+              icon="fas fa-file-medical-alt"
+              text="E. Expected Death Due to Illness or Pre-Existing Condition"
+              padding="0.6rem"
+            />
+          </Heading>
+
+          <ReportContainer>
+            <Render>
+              <Loader background="#3a3a40" />
+            </Render>
+          </ReportContainer>
+        </Section>
+      ) : sectionData && sectionData.length > 0 ? (
+        sectionRender
+      ) : (
+        <Section>
+          <Heading>
+            <HeadingPrimary
+              icon="fas fa-file-medical-alt"
+              text="E. Expected Death Due to Illness or Pre-Existing Condition"
+              padding="0.6rem"
+            />
+          </Heading>
+
+          <ReportContainer>
+            <Render>
+              <FieldData data="There is no Conditions Unequivocally Associated... data for this Patient" />
+            </Render>
+          </ReportContainer>
+        </Section>
+      )}
+    </>
+  );
+}
