+++ src/pages/subPages/DiagnosisOfDeath/renders/index.js
@@ -1,0 +1,20 @@
+// Export: DODFirst
+export { default as DODFirst } from "./DODFirst/DODFirst.component";
+
+// Export: DODSecond
+export { default as DODSecond } from "./DODSecond/DODSecond.component";
+
+// Export: DODThird
+export { default as DODThird } from "./DODThird/DODThird.component";
+
+// Export: DODFourth
+export { default as DODFourth } from "./DODFourth/DODFourth.component";
+
+// Export: DODFifth
+export { default as DODFifth } from "./DODFifth/DODFifth.component";
+
+// Export: DODSixth
+export { default as DODSixth } from "./DODSixth/DODSixth.component";
+
+// Export: DODSeventh
+export { default as DODSeventh } from "./DODSeventh/DODSeventh.component";
