+++ src/pages/subPages/PatientReport/renders/index.js
@@ -1,0 +1,2 @@
+// Export: PRFirst
+export { default as PRFirst } from "./PRFirst/PRFirst.component";
