--- src/pages/subPages/renders/index.js
@@ -1,2 +1,0 @@
-// Export: IncidentInformation
-export { default as IncidentInformation } from "./IncidentInformation/IncidentInformation.component";
